absorbed auto_qc into conversion_tools; added docstrings
